Conflicts:
	stack/corepersistence/collection/src/main/java/org/apache/usergrid/persistence/collection/util/EntityBuilder.java
	stack/corepersistence/graph/src/main/java/org/apache/usergrid/persistence/graph/impl/GraphManagerImpl.java
	stack/corepersistence/queryindex/src/main/java/org/apache/usergrid/persistence/index/utils/EntityBuilder.java
	stack/corepersistence/queryindex/src/main/java/org/apache/usergrid/persistence/index/utils/EntityMapUtils.java
	stack/corepersistence/queryindex/src/test/java/org/apache/usergrid/persistence/index/impl/EntityConnectionIndexImplTest.java
	stack/corepersistence/queryindex/src/test/java/org/apache/usergrid/persistence/index/impl/EntityIndexTest.java

+++ stack/corepersistence/collection/src/test/java/org/apache/usergrid/persistence/collection/cassandra/CassandraRule.java
@@ -59,6 +59,11 @@ public class CassandraRule extends EnvironResource {
 
         synchronized ( mutex ) {
 
+            //second into mutex
+            if(started){
+                return;
+            }
+
             File dataDir = Files.createTempDir();
             dataDir.deleteOnExit();
